LayoutTests/inspector/extensions/extensions-audits-api-expected.txt

Tests audits support in WebInspector Extensions API

Started extension.
Running tests...
RUNNING TEST: extension_testAuditsAPI
Added audit category, result dump follows:
{
    onAuditStarted : {
        addListener : <function>
        removeListener : <function>
    }
}
category.onAuditStarted fired, results dump follows:
{
    createURL : <function>
    createSnippet : <function>
    createText : <function>
    createObject : <function>
    createNode : <function>
    addResult : <function>
    createResult : <function>
    done : <function>
    Severity : {
        Info : "info"
        Warning : "warning"
        Severe : "severe"
    }
    createResourceLink : <function>
}
{
    contents : {
        0 : "Subtree"
    }
    children : {
    }
    expanded : false
    addChild : <function>
}
  Extension audits
All tests done.
